enable register_plist_parser

+++ cocos/2d/CCTMXXMLParser.cpp
@@ -784,7 +784,7 @@ void TMXMapInfo::endElement(void* /*ctx*/, const char *name)
     }
 }
 
-void TMXMapInfo::textHandler(void* /*ctx*/, const char *ch, size_t len)
+void TMXMapInfo::textHandler(void* /*ctx*/, const char *ch, int len)
 {
     TMXMapInfo *tmxMapInfo = this;
     std::string text(ch, 0, len);
