+++ cocos/platform/CCFileUtils.cpp
@@ -289,7 +289,7 @@ public:
         _state = SAX_NONE;
     }
 
-    void textHandler(void *ctx, const char *ch, size_t len) override
+    void textHandler(void *ctx, const char *ch, int len) override
     {
         if (_state == SAX_NONE)
         {
